requirements.txt
certifi==2016.8.8
chardet==2.3.0
click==6.6
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
Flask==0.11.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
requests==2.11.1
static3==0.7.0
Werkzeug==0.11.10
whitenoise==3.2
model-mommy==1.2.6
Pillow
psycopg2
djangorestframework
Markdown
django-filter
django-cors-headers
django-rest-swagger
pygments
easy-thumbnails==2.4.1
django-sortedm2m==1.4.0
